Add stronger teacher (GCNII) experiment and feature space analysis
Key additions:
- GCNII teacher experiment: Student still beats SOTA teacher by +1.39%
- Feature space analysis: Student has 23.6% better Davies-Bouldin Index
- Publication-quality figures (Figure 1-3) in PDF and PNG formats
- Updated README.md and RESULTS.md with comprehensive findings

Results summary:
- GCNII Teacher: 33.91%, Student: 35.30% (gap: +1.39%)
- Davies-Bouldin: Teacher 18.35 vs Student 14.01 (23.6% improvement)
- Silhouette Score: Teacher -0.038 vs Student -0.013 (65.8% improvement)
- Compactness Ratio: Teacher 4.99 vs Student 3.92 (21.4% improvement)

+++ analysis/feature_visualization.py
@@ -288,8 +288,8 @@ def run_feature_analysis(dataset='actor', num_runs=3, device='cuda'):
         for key in metrics_list[0].keys():
             values = [m[key] for m in metrics_list]
             avg_metrics[key] = {
-                'mean': np.mean(values),
-                'std': np.std(values)
+                'mean': float(np.mean(values)),
+                'std': float(np.std(values))
             }
         
         summary[model_name] = avg_metrics
